requirements.txt
# Core LLM + LangChain
openai
langchain
langchain-openai
langchain_community

# Vector DB
qdrant-client

# Utilities
python-dotenv
tqdm
tiktoken
jinja2
